TestExpectSuccess support for nil concrete types
When a nil variable of concrete type is passed to a function taking an
interface, the interface variable is non-nil.  Given variables of
concrete types implementing `error`, you can have a nil variable which
results in a non-nil `error` parameter.

Add tests for `TestExpectSuccess()`; the 4th test fails without the
corresponding change to functionality.

Change `TestExpectSuccess()` to use a reflect-based helper which pokes
into the interface variable and determines if the underlying object is
actually nil.

This changes the contract of the `TestExpectSuccess()` function, such
that a variable of concrete type can be meaningfully used as the error
parameter.  Since the only change is to types which could not
meaningfully be tested before, as they would always fail the test, this
seems reasonable.