Implemented basic deep merge functionality
Merge functionality currently focuses on recursively merging maps, where any
entries that exist in both maps and are a map will resurse to merge. All other
values will default to values in src overwriting dst.

This implementation's primary use case is around handling merges that were
simply JSON unmarshaled to a map[string]interface{} and not around structs or
maps of other base Go types.

It currently doesn't do anything special for slices, so a slice in src will
overwrite a slice in dst. Attempting to merge can be difficult, since there is
no way to overwrite an element in a map that is in a slice. Index evaluations
will never be successful. There also isn't an immediate need for merging a
slice.