Remove the backported archive/tar in tarhelper
We had pulled in the Go 1.2 version of archive/tar to resolve an issue that we encountered before 1.2 was official released, however after we moved to 1.2, we never removed the archive/tar we'd pulled in.

This commit just removes the vendored archive/tar and changes back to using the standard library.