Auto-create directory paths under the filesystem for files
Tars normally reference the directory a file is in before referencing files in
that directory, but this isn't always the case.

Darwin has "tar" and "gnutar". "tar" doesn't always do the right thing. If you
do "tar -czf foo.tar foo", where "foo" is a directory with files in it, the raw
tar file does not have an entry for the "foo" directory, it starts right into
the files. This would cause an error with tarhelper.

This change will auto-create the directories a file references if the directory
doesn't already exist. By the point it does it, the file has already been
validated as being in a safe location (not trying to break out of the
VirtualPath).