Added support to tarhelper for uid/gid mapping with tar and untar
The mapping is used to give the caller an option of specifying the uid/gid to
use. This can be used for remapping the uid/gid on the filesystem and can also
be used for enforcement of allowed values.

It does call the function each time, rather than maintaining an internal map.
Part of the concern was this as an exploit vector for lots of files with lots of
ids, but also adds overhead when someone may just want to add an offset for the
ids, something quick.