Added Subtract method to the IPRangeAllocator
This allows carving out an overlapping IPRange within the allocator by making
all of its IPs as reserved. This allows some cleaner IP configuration in the
orchestrator by allowing herited ranges and carving out reservations that
override the base default.

This also adds a fix for the Reserve call which didn't verify that the
provided IP was in the range. This would cause entries to be added to the
reservation that weren't harmful, but it would also decrement the remaining
counter making it think it was filling up.