Fix issue with deepmerge where merging child hashes resulted in a copy
This fixes behavior in deepmerge where objects are simplied copied into the
destination hash, which can result in multiple merge calls modifying the
source hash.

This currently uses an ugly hack that is well prefaced with the comments, but
it JSON marshals/unmarshals it to ensure all objects are a separate copy. This
can result in modifications such as int to float64, and can result in limiting
object types such as only maps with string keys. However this is necessary to
not modify the source object with future updates on the destination, and
because of no true deepcopy support with Go.

Note that currently, deepmerge is currently only consumed by the Orchestrator,
and this change is to fix a bug in the Orchestrator where some sets in the
cluster configuration would not be effective when applied and it was because
the Orchestrator is merging multiple configuration portions and it would end
of modifying the lower layers, which get persisted and then re-applied in
later runs without accounting for the changes further up properly.