improve test diagnostics; pre-clean; test correctly
Use optional messages on testtool functions, to identify which files are
causing problems.

Use `t.Fatal()` rather than `panic()` in a couple of places.

Nuke `/var/tmp/yy` before the test starts, so that a previous failure
doesn't persist (but if someone else on the same box created it, then
scream).  Yes, the `/var/tmp` usage is non-isolation of tests, but not
much to do about that (short of constructing path-names even more
dynamically).

Ensure that our non-existence tests are testing the correct files.