symlinks: block escape, continue past prohibited
Check that a relative symlink does not escape from our extraction root.

Allow tarball extraction to continue past a prohibited item, as long as
the FS is okay and the tarball is structurally okay.

Note: the continuing means that we don't get an extraction error if one
item is not extracted because of a bad entry, which is an API change.
This approach is open to debate.