Fix and discuss a few issues
One issue was that in the last commit the 'empty' path sections were improperly handled
leading to paths ending in '/' to be treated as if they had empty sections.

Another issue was the use of the path.Match method.  The wildcard character will not
cross over path seperators so it was unable to really detect if a path was a child of the
path in question.  `String.HasPrefix` is better able to handle that simple comparison.

A lasting issue that I'm not sure how to resolve is that when a symbolic link fails to be
untared, any file trying to go into that path doesn't get handled in the originally intended
way.  My comments I added point out the test case I'm talking about; `./etc` which points
to an absolute path `/realetc` is not created due to escaping from the intended root.  However,
a file which tries to get written into that was originally intended to go into a path
`./realetc/' which is relative to the untar.target rather than the actual root.  However, it now
just tries to be written into the `./etc/` path and when it isn't found it is created as a
normal directory.