Windows Permissions
Windows does not support the idea of an 'executable' bit on permissions.
As a result, no file tar'd on Windows would be executable causing many apps
to fail.  This included simple demo ruby/bash apps.  This fix just causes
everything tar'd on a Windows machine to be marked as executable.

Ideally the fix would be a little more fine grained, but this seems to be
the only way to ensure that all files needed as +x are marked as such.  In
Windows, its almost as if all files are executable so this seems to be in
line with their approach.

Docker's Window CLI faced this same concern and so our code fix and approach
mirror theirs.  See https://github.com/docker/docker/issues/11047