Fix handling of umask around chmod operations.
This updates the wrappers for Windows compatibility on umask to return the old
umask.

This updates the mknod handling, which already was doing umask(0) to ensure it
resets the umask back after it is done.

This also updates the chmod and create operations operations done to regular
files/directories to be wrapped in a umask(0) operation and ensure the umask
is reset after the chmod.

The handling does not reset with a defer to ensure the umask is reset
immediately before any other file operations might be done.