Added tempfile package for a file that deletes itself on close.
Added the tempfile package, which can be used to take in a reader,
reads all of its content to a local temporary file, and handles removing
the temp file automatically on deletion.

The interface it has allows reading, seeking, and closing.

This allows the logic to handle opening a file, whether local or remote,
in one place of code, and the logic for using that data in another
place. This allows the opening logic to not need to clean it up, it will
self-cleanup.