Adds package for gathering a process's CPU and memory usage.
This adds a function which can be used to calcualte a process's CPU
and memory usage via the /proc filesystem. This is mostly intended for
use in common component, which currently shells out to ps to do this.

Additionally, this moves towards CPU usage being shown in cumulative
time rather than doing a percentage. Percentage is already doing an
average over the number of CPUs and implies some timewindow to it. The
raw time spent is a rawer metric.