tarhelper: Ensure auto-created target directory gets owners mapped
This fixes an issue which can happen if the first entry in a tar file is
a file entry within a directory, and not a tar entry to create the
directory. For instance, doing `tar foo.tar app/*`, there won't be an
entry for `app/` in the tar file.

Tarhelper would automatically create the destination path, however it
was using a MkdirAll, which could create multiple directory elements,
and then was chowning only the last element. Further, it was doing the
chown without calling to OwnerMappingFunc/GroupMappingFunc, which is
what we use to apply user namespace offsets.

This would mean within the container, the directory path would exist but
with owners the container couldn't access.

Fixes #ENGT-6060