envmap: Fix envmap to be able to be JSON marshalled
This fixes EnvMap to have exported fields so that it can be properly
marshalled/unmarshalled into JSON.

This fixes an issue where internally, we were attempting to marshal an
EnvMap to cache its state on disk. However, it was being written as an
empty map.

This also adds a simple marshal/unmarshal test and validate the map
resolves the same as before.