docker: store user info in *url.Userinfo instead of string
Currently the userinfo being placed into a string meant that it would
get confusing how to use it since it could be a percent-encoded string (for
URLs), or a non-encoded string (for use in other method calls requiring auth),
or even a token which is already base64 encoded.

This PR makes the user info actually get stored into a Userinfo object. If the
user has a token, they can simply base64 decode it and store it in the Userinfo
field as normal. This ensures that it is clear how the data is to be used since
it is stored in a non-ambiguous format.

Since url.Userinfo has no exported fields, it normally would not marshal to JSON
correctly so custom Marshal and Unmarshal methods were added along with tests.