tarhelper: Add support for custom entry handlers on tar/untar
This adds support for a CustomHandler slice onto Tar and Untar. This
allows the caller to pass in function handlers that can be called to
customize the handling around specific entries.

Support to inject custom transformations is necessary to handle inline
conversion handlers for multiple union filesystems in the Continuum
codebase. We'll have this work by on the fly adapting to/from what we'll
officially utilize in the base packages to what the local host is
configured to use.

Change is part of #ENGT-9504.